requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.2
pyyaml>=5.4.1
matplotlib>=3.3.4
pillow>=8.1.0
scipy>=1.6.1
tqdm>=4.58.0
tensorboard>=2.4.1
pytorch-fid>=0.2.0
wandb>=0.10.20
pandas>=1.2.3
scikit-learn>=0.24.1
lpips>=0.1.4